requirements_tester.txt
telethon>=1.36.0
paramiko>=3.0.0
watchdog>=4.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
rich>=13.0.0
